# perf(bencode): use sync.Map for struct field caching to reduce contention