--- a/bencode/encode_test.go
+++ b/bencode/encode_test.go
@@ -91,15 +91,15 @@ func TestRandomEncode(t *testing.T) {
 	}
 }
 func BenchmarkGetEncodeFields(b *testing.B) {
-    type TestStruct struct {
-        Field1 string `bencode:"field1"`
-        Field2 int    `bencode:"field2"`
-    }
-    testType := reflect.TypeOf(TestStruct{})
-    b.ResetTimer()
-    b.RunParallel(func(pb *testing.PB) {
-        for pb.Next() {
-            getEncodeFields(testType)
-        }
-    })
+	type TestStruct struct {
+		Field1 string `bencode:"field1"`
+		Field2 int    `bencode:"field2"`
+	}
+	testType := reflect.TypeOf(TestStruct{})
+	b.ResetTimer()
+	b.RunParallel(func(pb *testing.PB) {
+		for pb.Next() {
+			getEncodeFields(testType)
+		}
+	})
 }